// paintingProject/src/app/ComponentLibrary/MyHeading/myHeading.types.ts
export enum HeadingType {
    H1 = "h1",
    H2 = "h2",
    H3 = "h3",
    H4 = "h4",
    H5 = "h5",
    H6 = "h6",
}

export enum HeadingColor {
    DARK = "dark",
    LIGHT = "light",
    INHERIT = "inherit",
}

export enum HeadingWeight {
    BOLD = "bold",
    NORMAL = "normal",
    INHERIT = "inherit",
}